more navbar css
