+++ frontend/src/stylesheets/app.scss
@@ -1,5 +1,5 @@
 body{
-    background-color: rgb(34, 82, 52);
+    background-color: rgb(11, 11, 11);
     display: block;
     position: relative;
     text-decoration: none;
