+++ frontend/src/stylesheets/login.scss
@@ -62,7 +62,7 @@
     background-color: seagreen;
     color: white;
     padding: 8px 24px;
-    border-radius: 3px;
+    border-radius: 7px;
     font-weight: 600;
     width: 45%;
     font-size: 1rem;
@@ -85,7 +85,7 @@
     background-color: rgb(89, 1, 152);
     color: white;
     padding: 8px 24px;
-    border-radius: 3px;
+    border-radius: 7px;
     font-weight: 600;
     width: 50%;
     font-size: 1rem;
