+++ frontend/src/stylesheets/main.scss
@@ -1,23 +1,29 @@
 .main-page-div {
-  display: inline-flex;
-  flex-direction: row;
+  background-color: rgb(7, 120, 181);
+  position: absolute;
+  display: flex;
+  // flex-direction: row;
   text-align: center;
   justify-content: center;
+  align-items: center;
   width: 100%;
-  height: 100%;
+  height: 100vh;
 }
 
 .container {
-  position: relative;
-  width: 1000px;
-  height: 500px;
-  margin: 240px auto;
-  background: blue;
-  justify-content: space-around;
+  display: flex;
+  // flex-direction: column;
+  // position: relative;
+  // left: 25%;
+  // width: 1000px;
+  height: 55%;
+  background: rgb(96, 96, 96);
+  // justify-content: space-around;
+  margin-bottom: 5rem;
 }
 
 .container .box {
-  position: relative;
+  // position: relative;
   width: (500px - 20px);
   height: (500px - 20px);
   float: left;
