+++ frontend/src/stylesheets/navbar.scss
@@ -1,51 +1,51 @@
-.navbar-div {
-    z-index: 2;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-around;
-    align-items: center;
-    // border-bottom: 1px solid #e6e6e6;
-    background-color: black;
-    height: 88px;
-    position: fixed;
-    top: 0;
-    width: 100%;
-    transition: 0.2s;
+.navbar {
+    z-index: 1;
+    height: 75px;
+}
+
+.nav-content-container{
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.nav-home-link{
+    background-color: rgb(16, 84, 168);
     color: white;
+    padding: 8px 20px 8px 24px;
+    right: 20px;
+    border-radius: 7px;
+    border: 1px solid dimgray;
+    text-decoration: none;
+    font-size: 1.5rem;
+    font-family: Lora;
+    margin-left: 5%;
+}
+
+.nav-home-link:hover{
+  background-color: rgb(44, 116, 204);
 }
 
 .nav-header{
   font-family: Lora;
-  font-size: 20px;
-  color: lightgray
+  font-size: 2rem;
+  color: lightgray;
+  font-weight: bold;
 }
 
-.dropdown {
+.nav-dropdown-button {
   position: relative;
   display: inline-block;
   cursor: pointer;
-//   padding-left: 
-}
-
-.dropdown-content {
-  display: none;
-  position: absolute;
-  background-color: gray;
-  min-width: 160px;
-  box-shadow: 0px 8px 16px 0px rgba(0,0,0,0.2);
-  padding: 12px 16px;
-  z-index: 1;
-  top: 5px;
-  right: 0px;
-  border-radius: 10px;
-}
-
-.dropdown:hover .dropdown-content {
-  display: block;
+  height: 64px;
+  margin-right: 5%;
 }
 
 .nav-icon{
-  z-index: 100;
+  z-index: 1;
+  height: 100%;
 }
 
 .nav-icon:hover {
@@ -57,22 +57,39 @@
     @keyframes spin { 100% { -webkit-transform: rotate(360deg); transform:rotate(360deg); } }
 }
 
-.options-list a{
+.dropdown-menu {
+  display: none;
+  position: absolute;
+  background-color: gray;
+  min-width: 160px;
+  box-shadow: 0px 8px 16px 0px rgba(0,0,0,0.2);
+  padding: 12px 16px;
+  z-index: 2;
+  right: 5%;
+  top: 70px;
+  border-radius: 7px;
+}
+
+.show{
+  display: block;
+}
+
+.options-list a, p{
+    display: flex;
+    background-color: rgb(16, 84, 168);
     list-style-type: none;
     text-decoration: none;
     color: white;
-    height: 100%;
-    width: 100%;
-    
+    padding: 8px   20px;
+    margin: 1em 0;
+    border-radius: 7px;
+    border: 1px solid dimgray;
+    font-size: 0.5rem;
+    font-family: Lora;
+    text-align: center;
+    justify-content: center;
 }
 
-.nav-home-link{
-  color: lightgray;
-  text-decoration: none;
-  font-size: 1.5rem;
-  font-family: Lora;
-}
-
-.nav-home-link:hover{
-  color: rgb(67, 97, 196);
+.options-list a:hover, p:hover{
+  background-color: rgb(44, 116, 204);
 }
