+++ frontend/src/stylesheets/signup.scss
@@ -67,7 +67,7 @@
     background-color: seagreen;
     color: white;
     padding: 8px 24px;
-    border-radius: 3px;
+    border-radius: 7px;
     font-weight: 600;
     width: 45%;
     font-size: 1rem;
